chore: make naming in rosetta conversion consistent
